feat(api): add methods to free DMA descriptors
The API prior this change only supports deallocating DMA channels (via method `(void)Adafruit_ZeroDMA::free(void)`), but it did not and **could not** deallocate the descriptor linked lists written to them. An _optional_ parameter was added to this method to also free the descriptor chain — but not by default — to retain backwards-compatibility.